feat(project): Use API to retrieve data
This commit changes the way projects are retrieved by using an API call instead of a mock. The `CommonService` now has a new method called `getProjectIci()` that retrieves the projects from the API and sets them in the component. The `HomeComponent` and `ProjectsComponent` have been updated to use this new method. Additionally, the project model has been imported into the HomeComponent and used to type check some variables. Finally, a routerlink attribute was added to a button in HomeComponent's HTML template so that it navigates to ProjectsComponent when clicked.