fix(user): Fix registration process and add image upload functionality
The code changes in this commit fix the registration process by separating the registration with and without an image. Previously, the code attempted to append a file to the FormData regardless of whether a file was selected or not. Now, if no file is selected, the registerData object is used for registration. If a file is selected, the registerImageData object is used instead.

Additionally, a new method `registerImage` has been added to the `AuthService` class to handle registration with an image.

In the `ProfilComponent`, there is a comment indicating that a modal angular should be implemented in the `openDialog` method.

These changes aim to improve user experience during registration and provide support for uploading profile images.